requirement.txt

streamlit
python-docx
python-pptx
pdfplumber
pandas
beautifulsoup4
google-generativeai
python-dotenv